@@ -1,6 +1,6 @@
 {
 	"name": "svg-chunk-webpack-plugin",
-	"version": "1.0.0",
+	"version": "2.0.0",
 	"description": "Generate SVG sprites according to entrypoint dependencies. Each page only imports its own svgs, wrapped as a sprite and optimized by svgo",
 	"keywords": [
 		"svg",
@@ -32,7 +32,7 @@
 		"build:example": "rm -rf ./example/dist/ && webpack --config=./example/webpack.config.js --mode=production",
 		"build:tsc": "tsc",
 		"coverage": "npm run test:coverage && cat ./coverage/lcov.info | coveralls",
-		"inspect:example": "rm -rf ./example/dist/ && node --inspect node_modules/webpack/bin/webpack.js --config=./example/webpack.config.js --mode=development --hidden=node_modules/",
+		"inspect:example": "rm -rf ./example/dist/ && node --inspect node_modules/webpack/bin/webpack.js --config=./example/webpack.config.js --mode=development",
 		"start:example": "rm -rf ./example/dist/ && webpack --config=./example/webpack.config.js --mode=development",
 		"test": "jest --verbose --coverage",
 		"test:coverage": "jest --coverage",
@@ -57,13 +57,19 @@
 	},
 	"prettier": {
 		"printWidth": 100,
+		"useTabs": true,
+		"semi": true,
+		"trailingComma": "none",
+		"singleQuote": true,
+		"arrowParens": "always",
 		"overrides": [
 			{
 				"files": "*.md",
 				"options": {
 					"proseWrap": "preserve",
 					"tabWidth": 2,
-					"useTabs": false
+					"useTabs": false,
+					"singleQuote": true
 				}
 			}
 		]
@@ -100,6 +106,14 @@
 			"semi": [
 				1,
 				"always"
+			],
+			"space-before-function-paren": [
+				"error",
+				{
+					"anonymous": "never",
+					"named": "never",
+					"asyncArrow": "always"
+				}
 			]
 		},
 		"globals": {
@@ -121,46 +135,45 @@
 	},
 	"dependencies": {
 		"extend": "^3.0.2",
-		"fs-extra": "^8.1.0",
 		"svgo": "^1.3.2",
 		"svgstore": "^3.0.0-2"
 	},
 	"devDependencies": {
-		"@babel/core": "^7.9.0",
-		"@babel/preset-env": "^7.9.0",
-		"@babel/preset-typescript": "^7.9.0",
-		"@types/fs-extra": "^8.1.0",
-		"@types/jest": "^25.2.1",
-		"@types/node": "^13.13.2",
-		"@types/tapable": "^1.0.5",
-		"@types/webpack": "^4.41.12",
+		"@babel/core": "^7.12.3",
+		"@babel/preset-env": "^7.12.1",
+		"@babel/preset-typescript": "^7.12.1",
+		"@types/fs-extra": "^9.0.3",
+		"@types/jest": "^26.0.15",
+		"@types/node": "^14.14.6",
+		"@types/tapable": "^1.0.6",
+		"@types/webpack": "^4.41.24",
 		"babel-eslint": "^10.1.0",
 		"babel-loader": "^8.1.0",
 		"coveralls": "^3.0.11",
-		"css-loader": "^3.4.2",
-		"eslint": "^6.8.0",
-		"eslint-config-standard": "^14.1.1",
-		"eslint-plugin-import": "^2.20.1",
+		"css-loader": "^5.0.1",
+		"css-minimizer-webpack-plugin": "^1.1.5",
+		"eslint": "^7.13.0",
+		"eslint-config-standard": "^16.0.1",
+		"eslint-plugin-import": "^2.22.1",
 		"eslint-plugin-node": "^11.0.0",
 		"eslint-plugin-promise": "^4.2.1",
-		"eslint-plugin-standard": "^4.0.1",
-		"jest": "^25.2.3",
-		"markdownlint-cli": "^0.22.0",
-		"mini-css-extract-plugin": "^0.9.0",
-		"optimize-css-assets-webpack-plugin": "^5.0.3",
-		"prettier-eslint": "^9.0.1",
+		"eslint-plugin-standard": "^4.0.2",
+		"jest": "^26.6.3",
+		"markdownlint-cli": "^0.24.0",
+		"mini-css-extract-plugin": "^1.3.0",
+		"optimize-css-assets-webpack-plugin": "^5.0.4",
+		"prettier-eslint": "^11.0.0",
 		"svgo-loader": "^2.2.1",
-		"terser-webpack-plugin": "^2.3.5",
-		"ts-jest": "^25.4.0",
-		"typescript": "^3.8.3",
-		"webpack": "^4.42.1",
-		"webpack-cli": "^3.3.11",
-		"webpack-manifest-plugin": "^2.2.0"
+		"terser-webpack-plugin": "^5.0.3",
+		"ts-jest": "^26.4.4",
+		"typescript": "^4.0.5",
+		"webpack": "^5.10.3",
+		"webpack-cli": "^4.2.0"
 	},
 	"peerDependencies": {
-		"webpack": "^4.0.0"
+		"webpack": ">= 5.10.3"
 	},
 	"engines": {
-		"node": ">=8.11.2"
+		"node": ">= 10.13.0"
 	}
 }
