@@ -1,25 +1,24 @@
 /**
  * @license MIT
  * @name SvgChunkWebpackPlugin
- * @version 1.0.0
+ * @version 2.0.0
  * @author: Yoriiis aka Joris DANIEL <joris.daniel@gmail.com>
  * @copyright 2020 Joris DANIEL
  **/
 import { Compiler } from 'webpack';
-import { Svgs, SpriteManifest, Sprites } from './interfaces';
+import { Svgs, SpriteManifest, Sprites, NormalModule } from './interfaces';
 declare class SvgSprite {
     options: {
+        filename: string;
         svgstoreConfig: Object;
         svgoConfig: Object;
         generateSpritesManifest: Boolean;
         generateSpritesPreview: Boolean;
-        filename: string;
     };
     svgOptimizer: any;
     spritesManifest: SpriteManifest;
     spritesList: Array<Sprites>;
     compilation: any;
-    entryNames: Array<string>;
     PLUGIN_NAME: any;
     /**
      * Instanciate the constructor
@@ -36,10 +35,13 @@ declare class SvgSprite {
      * Hook expose by the Webpack compiler
      *
      * @param {Object} compilation The Webpack compilation variable
-     *
-     * @returns {Promise<void>} Resolve the promise when all actions are done
      */
-    hookCallback(compilation: object): Promise<void>;
+    hookCallback(compilation: Object): Promise<void>;
+    /**
+     * Add assets
+     * The hook is triggered by webpack
+     */
+    addAssets(): Promise<void>;
     /**
      * Get entrypoint names from the compilation
      *
@@ -53,7 +55,7 @@ declare class SvgSprite {
      *
      * @returns {Promise<void>} Resolve the promise when all actions are done
      */
-    processEntry: (entryName: string) => Promise<void>;
+    processEntry(entryName: string): Promise<void>;
     /**
      * Optimize SVG with Svgo
      *
@@ -61,7 +63,7 @@ declare class SvgSprite {
      *
      * @returns {Promise<Svgs>} Svg name and optimized content with Svgo
      */
-    optimizeSvg: (filepath: string) => Promise<Svgs>;
+    optimizeSvg: (moduleDependency: NormalModule) => Promise<Svgs>;
     /**
      * Get SVGs filtered by entrypoints
      *
@@ -69,7 +71,7 @@ declare class SvgSprite {
      *
      * @returns {Array<Object>} Svgs list
      */
-    getSvgsByEntrypoint(entryName: string): Array<string>;
+    getSvgsDependenciesByEntrypoint(entryName: string): Array<NormalModule>;
     /**
      * Generate the SVG sprite with Svgstore
      *
@@ -131,18 +133,5 @@ declare class SvgSprite {
      * Create sprites preview
      */
     createSpritesPreview(): void;
-    /**
-     * Get preview template
-     *
-     * @returns {String} Template for the preview
-     */
-    getPreviewTemplate(): string;
-    /**
-     * Get sprites list
-     * The list is used to create the preview
-     *
-     * @returns {Array<Sprites>} Sprites list
-     */
-    getSpritesList(): Array<Sprites>;
 }
 export = SvgSprite;
