@@ -2,13 +2,15 @@
 /**
  * @license MIT
  * @name SvgChunkWebpackPlugin
- * @version 1.0.0
+ * @version 2.0.0
  * @author: Yoriiis aka Joris DANIEL <joris.daniel@gmail.com>
  * @copyright 2020 Joris DANIEL
  **/
+const webpack = require('webpack');
+// https://github.com/webpack/webpack/issues/11425#issuecomment-686607633
+const { RawSource } = webpack.sources;
 const { util } = require('webpack');
-const fse = require('fs-extra');
-const path = require('path');
+const path = require("path");
 const svgstore = require('svgstore');
 const Svgo = require('svgo');
 const extend = require('extend');
@@ -27,44 +29,22 @@ class SvgSprite {
         // This need to find plugin from loader context
         this.PLUGIN_NAME = PACKAGE_NAME;
         /**
-         * Process for each entry
-    
-         * @param {String} entryName Entrypoint name
-         *
-         * @returns {Promise<void>} Resolve the promise when all actions are done
-         */
-        this.processEntry = async (entryName) => {
-            return new Promise(async (resolve) => {
-                const svgs = this.getSvgsByEntrypoint(entryName);
-                const svgsOptimized = await Promise.all(svgs.map(filepath => this.optimizeSvg(filepath)));
-                const sprite = this.generateSprite(svgsOptimized);
-                this.createSpriteAsset({ entryName, sprite });
-                // Update sprites manifest
-                this.spritesManifest[entryName] = svgs.map(filepath => path.relative(this.compilation.options.context, filepath));
-                this.spritesList.push({
-                    name: entryName,
-                    content: sprite,
-                    svgs: svgs.map(filepath => path.basename(filepath, '.svg'))
-                });
-                resolve();
-            });
-        };
-        /**
          * Optimize SVG with Svgo
          *
          * @param {String} filepath SVG filepath from Webpack compilation
          *
          * @returns {Promise<Svgs>} Svg name and optimized content with Svgo
          */
-        this.optimizeSvg = async (filepath) => {
-            const svgContent = await fse.readFile(filepath, 'utf8');
-            const svgOptimized = await this.svgOptimizer.optimize(svgContent);
+        this.optimizeSvg = async (moduleDependency) => {
+            const source = JSON.parse(moduleDependency.originalSource()._value);
+            const svgOptimized = await this.svgOptimizer.optimize(source);
             return {
-                name: path.basename(filepath, '.svg'),
+                name: path.basename(moduleDependency.userRequest, '.svg'),
                 content: svgOptimized.data
             };
         };
         const DEFAULTS = {
+            filename: '[name].svg',
             svgstoreConfig: {
                 cleanDefs: false,
                 cleanSymbols: false,
@@ -84,13 +64,14 @@ class SvgSprite {
                 ]
             },
             generateSpritesManifest: false,
-            generateSpritesPreview: false,
-            filename: '[name].svg'
+            generateSpritesPreview: false
         };
         this.options = extend(true, DEFAULTS, options);
         this.svgOptimizer = new Svgo(this.options.svgoConfig);
         this.spritesManifest = {};
         this.spritesList = [];
+        this.hookCallback = this.hookCallback.bind(this);
+        this.addAssets = this.addAssets.bind(this);
     }
     /**
      * Apply function is automatically called by the Webpack main compiler
@@ -98,32 +79,37 @@ class SvgSprite {
      * @param {Object} compiler The Webpack compiler variable
      */
     apply(compiler) {
-        compiler.hooks.emit.tapPromise('SvgSprite', this.hookCallback.bind(this));
+        compiler.hooks.thisCompilation.tap('SvgSprite', this.hookCallback);
     }
     /**
      * Hook expose by the Webpack compiler
      *
      * @param {Object} compilation The Webpack compilation variable
-     *
-     * @returns {Promise<void>} Resolve the promise when all actions are done
      */
-    hookCallback(compilation) {
-        return new Promise(async (resolve) => {
-            // Reset value on every new process
-            this.spritesManifest = {};
-            this.spritesList = [];
-            this.compilation = compilation;
-            const isDev = this.compilation.options.mode === 'development';
-            const entryNames = this.getEntryNames();
-            await Promise.all(entryNames.map(async (entryName) => await this.processEntry(entryName)));
-            if (this.options.generateSpritesManifest && isDev) {
-                this.createSpritesManifest();
-            }
-            if (this.options.generateSpritesPreview && isDev) {
-                this.createSpritesPreview();
-            }
-            resolve();
-        });
+    async hookCallback(compilation) {
+        this.compilation = compilation;
+        // PROCESS_ASSETS_STAGE_ADDITIONAL: Add additional assets to the compilation
+        this.compilation.hooks.processAssets.tapPromise({
+            name: 'SvgChunkWebpackPlugin',
+            stage: webpack.Compilation.PROCESS_ASSETS_STAGE_ADDITIONAL
+        }, this.addAssets);
+    }
+    /**
+     * Add assets
+     * The hook is triggered by webpack
+     */
+    async addAssets() {
+        // Reset value on every new process
+        this.spritesManifest = {};
+        this.spritesList = [];
+        const entryNames = this.getEntryNames();
+        await Promise.all(entryNames.map(async (entryName) => await this.processEntry(entryName)));
+        if (this.options.generateSpritesManifest) {
+            this.createSpritesManifest();
+        }
+        if (this.options.generateSpritesPreview) {
+            this.createSpritesPreview();
+        }
     }
     /**
      * Get entrypoint names from the compilation
@@ -134,27 +120,42 @@ class SvgSprite {
         return Array.from(this.compilation.entrypoints.keys());
     }
     /**
+     * Process for each entry
+
+     * @param {String} entryName Entrypoint name
+     *
+     * @returns {Promise<void>} Resolve the promise when all actions are done
+     */
+    async processEntry(entryName) {
+        const svgsDependencies = this.getSvgsDependenciesByEntrypoint(entryName);
+        const svgsOptimized = await Promise.all(svgsDependencies.map((moduleDependency) => this.optimizeSvg(moduleDependency)));
+        const sprite = this.generateSprite(svgsOptimized);
+        this.createSpriteAsset({ entryName, sprite });
+        // Update sprites manifest
+        this.spritesManifest[entryName] = svgsDependencies.map((moduleDependency) => path.relative(this.compilation.options.context, moduleDependency.userRequest));
+        this.spritesList.push({
+            name: entryName,
+            content: sprite,
+            svgs: svgsDependencies.map((moduleDependency) => path.basename(moduleDependency.userRequest, '.svg'))
+        });
+    }
+    /**
      * Get SVGs filtered by entrypoints
      *
      * @param {String} entryName Entrypoint name
      *
      * @returns {Array<Object>} Svgs list
      */
-    getSvgsByEntrypoint(entryName) {
-        let listSvgs = [];
+    getSvgsDependenciesByEntrypoint(entryName) {
+        const listSvgsDependencies = [];
         this.compilation.entrypoints.get(entryName).chunks.forEach((chunk) => {
-            chunk.getModules().forEach((module) => {
-                module.buildInfo &&
-                    module.buildInfo.fileDependencies &&
-                    module.buildInfo.fileDependencies.forEach((filepath) => {
-                        const extension = path.extname(filepath).substr(1);
-                        if (extension === 'svg') {
-                            listSvgs.push(filepath);
-                        }
-                    });
-            });
+            for (const module of this.compilation.chunkGraph.getChunkModulesIterable(chunk)) {
+                if (module.buildInfo && module.buildInfo.SVG_CHUNK_WEBPACK_PLUGIN) {
+                    listSvgsDependencies.push(module);
+                }
+            }
         });
-        return listSvgs;
+        return listSvgsDependencies;
     }
     /**
      * Generate the SVG sprite with Svgstore
@@ -164,7 +165,7 @@ class SvgSprite {
      * @returns {String} Sprite string
      */
     generateSprite(svgsOptimized) {
-        let sprites = svgstore(this.options.svgstoreConfig);
+        const sprites = svgstore(this.options.svgstoreConfig);
         svgsOptimized.forEach((svg) => {
             sprites.add(svg.name, svg.content);
         });
@@ -181,10 +182,7 @@ class SvgSprite {
     createSpriteAsset({ entryName, sprite }) {
         const output = sprite;
         const filename = this.getFileName({ entryName, output });
-        this.compilation.assets[filename] = {
-            source: () => output,
-            size: () => output.length
-        };
+        this.compilation.emitAsset(filename, new RawSource(output, false));
     }
     /**
      * Get the filename for the asset compilation
@@ -243,10 +241,7 @@ class SvgSprite {
      */
     getContentHash(output) {
         const { hashFunction, hashDigest } = this.compilation.outputOptions;
-        return util
-            .createHash(hashFunction)
-            .update(output)
-            .digest(hashDigest);
+        return util.createHash(hashFunction).update(output).digest(hashDigest);
     }
     /**
      * Create sprite manifest with Webpack compilation
@@ -255,33 +250,13 @@ class SvgSprite {
      */
     createSpritesManifest() {
         const output = JSON.stringify(this.spritesManifest, null, 2);
-        this.compilation.assets['sprites-manifest.json'] = {
-            source: () => output,
-            size: () => output.length
-        };
+        this.compilation.emitAsset('sprites-manifest.json', new RawSource(output, false));
     }
     /**
      * Create sprites preview
      */
     createSpritesPreview() {
-        fse.outputFileSync(`${this.compilation.options.output.path}/sprites-preview.html`, this.getPreviewTemplate());
-    }
-    /**
-     * Get preview template
-     *
-     * @returns {String} Template for the preview
-     */
-    getPreviewTemplate() {
-        return templatePreview(this.getSpritesList());
-    }
-    /**
-     * Get sprites list
-     * The list is used to create the preview
-     *
-     * @returns {Array<Sprites>} Sprites list
-     */
-    getSpritesList() {
-        return this.spritesList;
+        this.compilation.emitAsset('sprites-preview.html', new RawSource(templatePreview(this.spritesList), false));
     }
 }
 // @ts-ignore
