@@ -4,7 +4,7 @@ import SvgChunkWebpackPlugin from '../index';
 
 import {
 	mockGetEntryNames,
-	mockGetSvgsByEntrypoint,
+	mockGetSvgsDependenciesByEntrypoint,
 	mockOptimizeSvg,
 	mockGenerateSprite,
 	mockGetBuildHash,
@@ -14,10 +14,13 @@ import {
 import templatePreview from '../preview';
 
 import path from 'path';
-import fse from 'fs-extra';
 import Svgo from 'svgo';
 
 jest.mock('../preview');
+
+const webpack = require('webpack');
+const { util } = require('webpack');
+const { RawSource } = webpack.sources;
 
 // Mock Webpack util to works with chaining function
 jest.mock('webpack', () => {
@@ -31,15 +34,19 @@ jest.mock('webpack', () => {
 		digest: mockDigest
 	};
 	return {
+		Compilation: {
+			PROCESS_ASSETS_STAGE_ADDITIONAL: ''
+		},
 		util: {
 			createHash: mockCreateHash.mockImplementation(() => utils),
 			update: mockUpdate.mockImplementation(() => utils),
 			digest: mockDigest.mockImplementation(() => 'a5934d97b38c748213317d7e5ffd31b6')
+		},
+		sources: {
+			RawSource: jest.fn()
 		}
 	};
 });
-
-const { util } = require('webpack');
 
 // Save reference to path.relative because the function is mock inside tests
 // The original function is return after each tests
@@ -62,22 +69,40 @@ const spritesFixture = {
 		'<svg aria-hidden="true" style="position: absolute; width: 0; height: 0; overflow: hidden;"><defs><linearGradient id="a"><stop offset="5%" stop-color="green"/><stop offset="95%" stop-color="gold"/></linearGradient></defs><symbol id="gradient"><rect fill="url(#a)" width="100%" height="100%"/></symbol><symbol id="video" viewBox="0 0 16 16"><path d="M8 0C3.6 0 0 3.6 0 8s3.6 8 8 8 8-3.6 8-8-3.6-8-8-8zm3.3 8.5l-4.5 3c-.1 0-.1.1-.2.1s-.2 0-.3-.1c-.2-.1-.3-.3-.3-.5V5c0-.2.1-.4.2-.5.2-.1.3-.1.5 0l4.5 3c.2.1.3.3.3.5s-.1.4-.2.5z" fill="#ff004b"/></symbol><symbol id="smiley-love" viewBox="0 0 48 48"><circle class="st0" cx="24" cy="24" r="24" fill="#fbd971"/><path class="st1" d="M24 41.1c-7.6 0-13.7-6.2-13.7-13.7 0-.6.5-1.1 1.1-1.1.6 0 1.1.5 1.1 1.1 0 6.3 5.1 11.4 11.4 11.4s11.4-5.1 11.4-11.4c0-.6.5-1.1 1.1-1.1.6 0 1.1.5 1.1 1.1.2 7.6-5.9 13.7-13.5 13.7z" fill="#d8b11a"/><path d="M14.3 12.2c.5-1.1 1.6-1.9 3-1.9 1.8 0 3.1 1.5 3.2 3.2 0 0 .1.4-.1 1.2-.3 1.1-.9 2-1.7 2.8l-4.4 3.8-4.3-3.8c-.8-.7-1.4-1.7-1.7-2.8-.2-.8-.1-1.2-.1-1.2.2-1.8 1.5-3.2 3.2-3.2 1.4 0 2.4.8 2.9 1.9z" fill="#e64c3c"/><path data-name="Calque 1-2-2" d="M33.6 12.2c.5-1.1 1.6-1.9 3-1.9 1.8 0 3.1 1.5 3.2 3.2 0 0 .1.4-.1 1.2-.3 1.1-.9 2-1.7 2.8l-4.4 3.8-4.3-3.8c-.8-.7-1.4-1.7-1.7-2.8-.2-.8-.1-1.2-.1-1.2.2-1.8 1.5-3.2 3.2-3.2 1.3 0 2.4.8 2.9 1.9z" fill="#e64c3c"/></symbol></svg>'
 };
 
-const svgsFilepath = [
-	'example/src/svgs/gradient.svg',
-	'example/src/svgs/video.svg',
-	'example/src/svgs/smiley-love.svg'
+const svgsDependencies = [
+	{
+		buildInfo: {
+			SVG_CHUNK_WEBPACK_PLUGIN: true
+		},
+		userRequest: '/Applications/svg-chunk-webpack-plugin/example/src/svgs/gradient.svg',
+		originalSource: () => ({
+			_value: JSON.stringify(svgsFixture.gradient)
+		})
+	},
+	{
+		buildInfo: {
+			SVG_CHUNK_WEBPACK_PLUGIN: true
+		},
+		userRequest: '/Applications/svg-chunk-webpack-plugin/example/src/svgs/video.svg',
+		originalSource: () => ({
+			_value: JSON.stringify(svgsFixture.video)
+		})
+	},
+	{
+		buildInfo: {
+			SVG_CHUNK_WEBPACK_PLUGIN: true
+		},
+		userRequest: '/Applications/svg-chunk-webpack-plugin/example/src/svgs/smiley-love.svg',
+		originalSource: () => ({
+			_value: JSON.stringify(svgsFixture['smiley-love'])
+		})
+	}
 ];
 const svgsSprite = [
 	{ name: 'gradient', content: svgsFixture.gradient },
 	{ name: 'video', content: svgsFixture.video },
 	{ name: 'smiley-love', content: svgsFixture['smiley-love'] }
 ];
-const spritesList = {
-	name: 'home',
-	content: spritesFixture.home,
-	svgs: ['gradient', 'video', 'smiley-love']
-};
-
 const options = {
 	generateSpritesManifest: true,
 	generateSpritesPreview: true
@@ -89,41 +114,14 @@ const entrypointsMap = new Map();
 entrypointsMap.set('home', {
 	chunks: [
 		{
-			hash: 'beb18939e5093045258b8d24a34dd844',
-			getModules: () => [
-				{
-					buildInfo: { fileDependencies: ['example/src/js/home.js'] }
-				},
-				{
-					buildInfo: { fileDependencies: ['example/src/svgs/gradient.svg'] }
-				},
-				{
-					buildInfo: { fileDependencies: ['example/src/svgs/video.svg'] }
-				},
-				{
-					buildInfo: { fileDependencies: ['example/src/js/component.js'] }
-				},
-				{
-					buildInfo: {
-						fileDependencies: ['example/src/svgs/smiley-love.svg']
-					}
-				}
-			]
+			hash: 'beb18939e5093045258b8d24a34dd844'
 		}
 	]
 });
 entrypointsMap.set('news', {
 	chunks: [
 		{
-			hash: 'beb18939e5093045258b8d24a34dd843',
-			getModules: () => [
-				{
-					buildInfo: {
-						fileDependencies: ['example/src/svgs/smiley-surprised.svg']
-					}
-				},
-				{ buildInfo: { fileDependencies: ['example/src/svgs/popcorn.svg'] } }
-			]
+			hash: 'beb18939e5093045258b8d24a34dd843'
 		}
 	]
 });
@@ -145,6 +143,15 @@ beforeEach(() => {
 		outputOptions: {
 			hashFunction: 'md4',
 			hashDigest: 'hex'
+		},
+		emitAsset: jest.fn(),
+		hooks: {
+			processAssets: {
+				tapPromise: jest.fn()
+			}
+		},
+		chunkGraph: {
+			getChunkModulesIterable: () => svgsDependencies
 		}
 	};
 
@@ -158,6 +165,7 @@ afterEach(() => {
 describe('SvgChunkWebpackPlugin constructor', () => {
 	it('Should initialize the constructor with custom options', () => {
 		expect(svgChunkWebpackPlugin.options).toMatchObject({
+			filename: '[name].svg',
 			svgstoreConfig: {
 				cleanDefs: false,
 				cleanSymbols: false,
@@ -177,8 +185,7 @@ describe('SvgChunkWebpackPlugin constructor', () => {
 				]
 			},
 			generateSpritesManifest: true,
-			generateSpritesPreview: true,
-			filename: '[name].svg'
+			generateSpritesPreview: true
 		});
 		expect(svgChunkWebpackPlugin.svgOptimizer).toBeInstanceOf(Svgo);
 		expect(svgChunkWebpackPlugin.spritesManifest).toEqual({});
@@ -217,29 +224,46 @@ describe('SvgChunkWebpackPlugin apply', () => {
 	it('Should call the apply function', () => {
 		const compilerWebpack = {
 			hooks: {
-				emit: {
-					tapPromise: () => {}
+				thisCompilation: {
+					tap: () => {}
 				}
 			}
 		};
-		compilerWebpack.hooks.emit.tapPromise = jest.fn();
+		compilerWebpack.hooks.thisCompilation.tap = jest.fn();
 
 		svgChunkWebpackPlugin.apply(compilerWebpack);
 
-		expect(compilerWebpack.hooks.emit.tapPromise).toHaveBeenCalled();
+		expect(compilerWebpack.hooks.thisCompilation.tap).toHaveBeenCalled();
 	});
 });
 
 describe('SvgChunkWebpackPlugin hookCallback', () => {
-	it('Should call the hookCallback function', async () => {
+	it('Should call the hookCallback function', () => {
+		svgChunkWebpackPlugin.hookCallback(compilationWebpack);
+
+		expect(
+			svgChunkWebpackPlugin.compilation.hooks.processAssets.tapPromise
+		).toHaveBeenCalledWith(
+			{
+				name: 'SvgChunkWebpackPlugin',
+				stage: webpack.Compilation.PROCESS_ASSETS_STAGE_ADDITIONAL
+			},
+			svgChunkWebpackPlugin.addAssets
+		);
+	});
+});
+
+describe('SvgChunkWebpackPlugin addAssets', () => {
+	it('Should call the addAssets function', async () => {
 		mockGetEntryNames(svgChunkWebpackPlugin, entryNames);
 		svgChunkWebpackPlugin.processEntry = jest.fn();
 		svgChunkWebpackPlugin.createSpritesManifest = jest.fn();
 		svgChunkWebpackPlugin.createSpritesPreview = jest.fn();
 
-		await svgChunkWebpackPlugin.hookCallback(compilationWebpack);
+		await svgChunkWebpackPlugin.addAssets();
 
-		expect(svgChunkWebpackPlugin.compilation).toEqual(compilationWebpack);
+		expect(svgChunkWebpackPlugin.spritesManifest).toStrictEqual({});
+		expect(svgChunkWebpackPlugin.spritesList).toStrictEqual([]);
 		expect(svgChunkWebpackPlugin.getEntryNames).toHaveBeenCalled();
 		expect(svgChunkWebpackPlugin.processEntry).toHaveBeenCalledTimes(2);
 		expect(svgChunkWebpackPlugin.processEntry).toHaveBeenCalledWith('home');
@@ -248,7 +272,7 @@ describe('SvgChunkWebpackPlugin hookCallback', () => {
 		expect(svgChunkWebpackPlugin.createSpritesPreview).toHaveBeenCalled();
 	});
 
-	it('Should call the hookCallback function without manifest and preview', async () => {
+	it('Should call the addAssets function without manifest and preview', async () => {
 		mockGetEntryNames(svgChunkWebpackPlugin, entryNames);
 		svgChunkWebpackPlugin.processEntry = jest.fn();
 		svgChunkWebpackPlugin.createSpritesManifest = jest.fn();
@@ -256,7 +280,7 @@ describe('SvgChunkWebpackPlugin hookCallback', () => {
 
 		svgChunkWebpackPlugin.options.generateSpritesManifest = false;
 		svgChunkWebpackPlugin.options.generateSpritesPreview = false;
-		await svgChunkWebpackPlugin.hookCallback(compilationWebpack);
+		await svgChunkWebpackPlugin.addAssets();
 
 		expect(svgChunkWebpackPlugin.createSpritesManifest).not.toHaveBeenCalled();
 		expect(svgChunkWebpackPlugin.createSpritesPreview).not.toHaveBeenCalled();
@@ -273,7 +297,7 @@ describe('SvgChunkWebpackPlugin getEntryNames', () => {
 
 describe('SvgChunkWebpackPlugin processEntry', () => {
 	it('Should call the processEntry function', async () => {
-		mockGetSvgsByEntrypoint(svgChunkWebpackPlugin, svgsFilepath);
+		mockGetSvgsDependenciesByEntrypoint(svgChunkWebpackPlugin, svgsDependencies);
 		mockOptimizeSvg(svgChunkWebpackPlugin, svgsFixture);
 		mockGenerateSprite(svgChunkWebpackPlugin, spritesFixture.home);
 		svgChunkWebpackPlugin.createSpriteAsset = jest.fn();
@@ -282,31 +306,30 @@ describe('SvgChunkWebpackPlugin processEntry', () => {
 		// The original function is return after the test
 		path.relative = jest
 			.fn()
-			.mockImplementation((context, filepath) => `${context}${filepath}`);
+			.mockImplementation(
+				(context, moduleDependency) => `${context}${moduleDependency.userRequest}`
+			);
 
 		svgChunkWebpackPlugin.compilation = compilationWebpack;
 		await svgChunkWebpackPlugin.processEntry('home');
 
 		const entry = 'home';
-		expect(svgChunkWebpackPlugin.getSvgsByEntrypoint).toHaveBeenCalledWith(entry);
+		expect(svgChunkWebpackPlugin.getSvgsDependenciesByEntrypoint).toHaveBeenCalledWith(entry);
 		expect(svgChunkWebpackPlugin.optimizeSvg).toHaveBeenCalledTimes(3);
-		expect(svgChunkWebpackPlugin.optimizeSvg).toHaveBeenCalledWith(
-			'example/src/svgs/gradient.svg'
-		);
-		expect(svgChunkWebpackPlugin.optimizeSvg).toHaveBeenCalledWith(
-			'example/src/svgs/video.svg'
-		);
-		expect(svgChunkWebpackPlugin.optimizeSvg).toHaveBeenCalledWith(
-			'example/src/svgs/smiley-love.svg'
-		);
+		expect(svgChunkWebpackPlugin.optimizeSvg).toHaveBeenCalledWith(svgsDependencies[0]);
+		expect(svgChunkWebpackPlugin.optimizeSvg).toHaveBeenCalledWith(svgsDependencies[1]);
+		expect(svgChunkWebpackPlugin.optimizeSvg).toHaveBeenCalledWith(svgsDependencies[2]);
 		expect(svgChunkWebpackPlugin.generateSprite).toHaveBeenCalledWith(svgsSprite);
 		expect(svgChunkWebpackPlugin.createSpriteAsset).toHaveBeenCalledWith({
 			entryName: entry,
 			sprite: spritesFixture.home
 		});
 		expect(svgChunkWebpackPlugin.spritesManifest).toEqual({
-			home: svgsFilepath.map(
-				filepath => `${svgChunkWebpackPlugin.compilation.options.context}${filepath}`
+			home: svgsDependencies.map((moduleDependency) =>
+				path.relative(
+					svgChunkWebpackPlugin.compilation.options.context,
+					moduleDependency.userRequest
+				)
 			)
 		});
 		expect(svgChunkWebpackPlugin.spritesList).toEqual([
@@ -321,14 +344,12 @@ describe('SvgChunkWebpackPlugin processEntry', () => {
 
 describe('SvgChunkWebpackPlugin optimizeSvg', () => {
 	it('Should call the optimizeSvg function', async () => {
-		fse.readFile = jest.fn().mockImplementation(() => svgsFixture.gradient);
 		svgChunkWebpackPlugin.svgOptimizer.optimize = jest
 			.fn()
 			.mockImplementation(() => ({ data: svgsFixture.gradient }));
 
-		const result = await svgChunkWebpackPlugin.optimizeSvg(svgsFilepath[0]);
+		const result = await svgChunkWebpackPlugin.optimizeSvg(svgsDependencies[0]);
 
-		expect(fse.readFile).toHaveBeenCalledWith(svgsFilepath[0], 'utf8');
 		expect(svgChunkWebpackPlugin.svgOptimizer.optimize).toHaveBeenCalledWith(
 			svgsFixture.gradient
 		);
@@ -339,13 +360,37 @@ describe('SvgChunkWebpackPlugin optimizeSvg', () => {
 	});
 });
 
-describe('SvgChunkWebpackPlugin getSvgsByEntrypoint', () => {
-	it('Should call the getSvgsByEntrypoint function', () => {
+describe('SvgChunkWebpackPlugin getSvgsDependenciesByEntrypoint', () => {
+	it('Should call the getSvgsDependenciesByEntrypoint function', () => {
 		svgChunkWebpackPlugin.compilation = compilationWebpack;
+		const homeChunks = svgChunkWebpackPlugin.compilation.entrypoints.get('home');
+		svgChunkWebpackPlugin.compilation.entrypoints.get = jest.fn().mockReturnValue(homeChunks);
+		svgChunkWebpackPlugin.compilation.chunkGraph.getChunkModulesIterable = jest
+			.fn()
+			.mockReturnValue(svgsDependencies);
 
-		const result = svgChunkWebpackPlugin.getSvgsByEntrypoint('home');
+		const result = svgChunkWebpackPlugin.getSvgsDependenciesByEntrypoint('home');
 
-		expect(result).toEqual(svgsFilepath);
+		expect(svgChunkWebpackPlugin.compilation.entrypoints.get).toHaveBeenCalledWith('home');
+		expect(
+			svgChunkWebpackPlugin.compilation.chunkGraph.getChunkModulesIterable
+		).toHaveBeenCalledWith(homeChunks.chunks[0]);
+		expect(result).toEqual(svgsDependencies);
+	});
+
+	it('Should call the getSvgsDependenciesByEntrypoint function with no flag on the last item', () => {
+		svgChunkWebpackPlugin.compilation = compilationWebpack;
+		svgChunkWebpackPlugin.compilation.chunkGraph.getChunkModulesIterable = jest
+			.fn()
+			.mockReturnValue([
+				{
+					buildInfo: {}
+				}
+			]);
+
+		const result = svgChunkWebpackPlugin.getSvgsDependenciesByEntrypoint('home');
+
+		expect(result).toEqual([]);
 	});
 });
 
@@ -366,14 +411,15 @@ describe('SvgChunkWebpackPlugin createSpriteAsset', () => {
 
 		svgChunkWebpackPlugin.createSpriteAsset({ entryName: 'home', sprite: output });
 
-		expect(svgChunkWebpackPlugin.compilation.assets).toEqual({
-			'home.svg': {
-				source: expect.any(Function),
-				size: expect.any(Function)
-			}
+		expect(svgChunkWebpackPlugin.getFileName).toHaveBeenCalledWith({
+			entryName: 'home',
+			output
 		});
-		expect(svgChunkWebpackPlugin.compilation.assets['home.svg'].source()).toBe(output);
-		expect(svgChunkWebpackPlugin.compilation.assets['home.svg'].size()).toBe(output.length);
+		expect(RawSource).toHaveBeenCalledWith(output, false);
+		expect(svgChunkWebpackPlugin.compilation.emitAsset).toHaveBeenCalledWith(
+			'home.svg',
+			new RawSource(output, false)
+		);
 	});
 });
 
@@ -491,6 +537,7 @@ describe('SvgChunkWebpackPlugin getContentHash', () => {
 
 describe('SvgChunkWebpackPlugin createSpritesManifest', () => {
 	it('Should call the createSpritesManifest function', () => {
+		jest.spyOn(JSON, 'stringify');
 		svgChunkWebpackPlugin.compilation = compilationWebpack;
 
 		svgChunkWebpackPlugin.spritesManifest = {
@@ -503,53 +550,29 @@ describe('SvgChunkWebpackPlugin createSpritesManifest', () => {
 		const output = JSON.stringify(svgChunkWebpackPlugin.spritesManifest, null, 2);
 		svgChunkWebpackPlugin.createSpritesManifest();
 
-		expect(svgChunkWebpackPlugin.compilation.assets).toEqual({
-			'sprites-manifest.json': {
-				source: expect.any(Function),
-				size: expect.any(Function)
-			}
-		});
-		expect(svgChunkWebpackPlugin.compilation.assets['sprites-manifest.json'].source()).toBe(
-			output
-		);
-		expect(svgChunkWebpackPlugin.compilation.assets['sprites-manifest.json'].size()).toBe(
-			output.length
+		expect(JSON.stringify).toHaveBeenCalledWith(svgChunkWebpackPlugin.spritesManifest, null, 2);
+		expect(RawSource).toHaveBeenCalledWith(output, false);
+		expect(svgChunkWebpackPlugin.compilation.emitAsset).toHaveBeenCalledWith(
+			'sprites-manifest.json',
+			new RawSource(output, false)
 		);
 	});
 });
 
 describe('SvgChunkWebpackPlugin createSpritesPreview', () => {
 	it('Should call the createSpritesPreview function', () => {
-		fse.outputFileSync = jest.fn();
-		svgChunkWebpackPlugin.getPreviewTemplate = jest.fn().mockImplementation(() => 'template');
-
 		svgChunkWebpackPlugin.compilation = compilationWebpack;
+
 		svgChunkWebpackPlugin.createSpritesPreview();
 
-		expect(svgChunkWebpackPlugin.getPreviewTemplate).toHaveBeenCalled();
-		expect(fse.outputFileSync).toHaveBeenCalledWith(
-			`${compilationWebpack.options.output.path}/sprites-preview.html`,
-			'template'
+		expect(templatePreview).toHaveBeenCalledWith(svgChunkWebpackPlugin.spritesList);
+		expect(RawSource).toHaveBeenCalledWith(
+			templatePreview(svgChunkWebpackPlugin.spritesList),
+			false
 		);
-	});
-});
-
-describe('SvgChunkWebpackPlugin getPreviewTemplate', () => {
-	it('Should call the getPreviewTemplate function', () => {
-		svgChunkWebpackPlugin.getSpritesList = jest.fn().mockImplementation(() => spritesList);
-
-		svgChunkWebpackPlugin.getPreviewTemplate();
-
-		expect(svgChunkWebpackPlugin.getSpritesList).toHaveBeenCalled();
-		expect(templatePreview).toHaveBeenCalledWith(spritesList);
-	});
-});
-
-describe('SvgChunkWebpackPlugin getSpritesList', () => {
-	it('Should call the getSpritesList function', () => {
-		svgChunkWebpackPlugin.spritesList = spritesList;
-		const result = svgChunkWebpackPlugin.getSpritesList();
-
-		expect(result).toEqual(spritesList);
+		expect(svgChunkWebpackPlugin.compilation.emitAsset).toHaveBeenCalledWith(
+			'sprites-preview.html',
+			new RawSource(templatePreview(svgChunkWebpackPlugin.spritesList), false)
+		);
 	});
 });
